backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai==1.12.0
pyupbit==0.2.33
ta==0.11.0
selenium==4.18.1
webdriver-manager==4.0.1
Pillow==10.2.0
youtube-transcript-api==0.6.2
plotly==5.18.0
schedule==1.2.1
sqlalchemy==2.0.27
alembic==1.13.1
mysqlclient==2.2.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.3
gunicorn==21.2.0
apscheduler==3.10.4 